codeintel: RFC 619 tiny code flattening
